requirements.txt
flask
python-dotenv
openai
langchain
langchain_core
langchain_community
langchain_openai
langchain_text_splitters
pypdf
faiss-cpu
sentence-transformers
docx2txt
torch